102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,17 +10,17 @@
 
 int main(void)
 {
-	long long int num1 = 1;
-	long long int num2 = 2;
+	int num1 = 1;
+	int num2 = 2;
 	long long int num3;
 	int i;
 
-	printf("%lu, %lu, ", num1, num2);
+	printf("%d, %d, ", num1, num2);
 
 	for (i = 0; i <= 48; i++)
 	{
 		num3 = num1 + num2;
-		printf("%lu, ", num3);
+		printf("%lld, ", num3);
 		num1 = num2;
 		num2 = num3;
 	}
